Default easing modes.
Is it easeElasticIn or easeElasticOut? I can never remember. And it’s tedious to
type “InOut” when I want something simple like linear easing. This commit
removes easeLinearIn, easeLinearOut and easeLinearInOut; they’re all the same,
so it’s just easeLinear now. The new aliases are as follows:

    easeQuad    ↦ easeQuadInOut
    easeCubic   ↦ easeCubicInOut
    easePoly    ↦ easePolyInOut
    easeSin     ↦ easeSinInOut
    easeExp     ↦ easeExpInOut
    easeCircle  ↦ easeCircleInOut
    easeBounce  ↦ easeBounceOut
    easeBack    ↦ easeBackInOut
    easeElastic ↦ easeElasticOut

Thus, the default is typically in-out, except bounce and elastic which use out.